Use both POT and nslc, weighted.
Use bins which evenly weight the POT (translated into mean number
of slices) and the actual number of slices.  This is very kludgy,
as indicated by the comments, but is still more physically
meaningful that just using one or the other.

Also if I just use POT, there is very little range to work
with.  I need pretty good statistics in each intensity bin
and that's just not how the data is.