Fix number of tracks for pileup background.
I had erroneously been using the number of events marked as
from the pileup background sample with i == 0 as a count
of tracks for that sample.  But the ntuple generator does
not waste space by writing the track information out for both
the signal and the background.  Of course, the number of tracks
is the same for these two cases, so make that explicit.