Bayesian systematics, combine regions.
* Reimplement systematics in stage three using a Bayesian approach
where we convolute the probability distribution with a two-sided
Gaussian and then do the integration, rather than doing the integration
and then adding systematics in quadrature.  This is probably more
accurate, although the systematics are very conservative in any case.
The systematic errors still don't use a full likelihood from the
previous stage, but I'm not sure it would be all that meaningful
if they did.

* Now that that's done, it is possible to combine the main detector
and muon catcher results using their likelihoods (which are the same
as their posterior distributions), which is more accurate than
adding the results in quadrature.