Split NC scales into xsec and cal/acc.
About half of the NC scale error in the standard numu analysis
is from cross sections and the other half from calibration/acceptance.
So that we can use the same pull term for the latter for each, but
separate ones for the former, separate the NC scales into
one each for cross sections and one more for the calibration/acceptance.
These are multiplied together internally to get the scale, but
penalty terms are applied directly to the parts.  (Much easier than
keeping two scales and using a covariance matrix.)